move team ITs to their own class and break them up into smaller tests
